+++ lib/index.js
@@ -1,1 +1,1 @@
-const requests = require('./requests')
+const events = require('./events')
